fix details according to @multun review
