+++ src/main/java/fr/sncf/osrd/train/TrainPath.java
@@ -114,21 +114,22 @@ public class TrainPath {
      * If the path position is higher than the fullPath length the function return null. */
     public TrackSectionLocation findLocation(double pathPosition) {
         for (var track : trackSectionPath) {
-            if (pathPosition <= track.length()) {
-                var location = track.getBeginPosition();
-                if (track.direction == EdgeDirection.START_TO_STOP)
-                    location += pathPosition;
-                else
-                    location -= pathPosition;
-                return new TrackSectionLocation(track.edge, location);
+            if (pathPosition > track.length()) {
+                pathPosition -= track.length();
+                continue;
             }
-            pathPosition -= track.length();
+
+            var location = track.getBeginPosition();
+            if (track.direction == EdgeDirection.START_TO_STOP)
+                location += pathPosition;
+            else
+                location -= pathPosition;
+            return new TrackSectionLocation(track.edge, location);
         }
 
-        if (pathPosition < 1) {
-            // We might reach this point with an epsilon left when looking for the end because of float inaccuracies
+        // We might reach this point with an epsilon left when looking for the end because of float inaccuracies
+        if (pathPosition < 1e-3)
             return getEndLocation();
-        }
 
         return null;
     }
